refactor(menu): split remaining menu flows
