+++ src/sf33rd/Source/Game/menu/menu.c
@@ -281,413 +281,6 @@ void Menu_Init(struct _TASK* task_ptr) {
     cpReadyTask(TASK_SAVER, Saver_Task);
 }
 
-static void initialize_mode_select(struct _TASK* task_ptr, const s16 loop_counter) {
-    s16 ix;
-
-    FadeOut(1, 0xFF, 8);
-    task_ptr->r_no[2] += 1;
-    task_ptr->timer = 5;
-    Mode_Type = MODE_ARCADE;
-    Present_Mode = 1;
-
-    if (task[TASK_ENTRY].condition != 1) {
-        E_No[0] = 1;
-        E_No[1] = 2;
-        E_No[2] = 2;
-        E_No[3] = 0;
-        cpReadyTask(TASK_ENTRY, Entry_Task);
-    }
-
-    Menu_Common_Init();
-
-    for (ix = 0; ix < 4; ix++) {
-        Menu_Suicide[ix] = 0;
-    }
-
-    Clear_Personal_Data(0);
-    Clear_Personal_Data(1);
-    Menu_Cursor_Y[0] = Cursor_Y_Pos[0][0];
-
-    if (Menu_Cursor_Y[0] >= loop_counter) {
-        Menu_Cursor_Y[0] = loop_counter - 1;
-    }
-
-    Cursor_Y_Pos[0][1] = 0;
-    Cursor_Y_Pos[0][2] = 0;
-    Cursor_Y_Pos[0][3] = 0;
-
-    for (ix = 0; ix < 4; ix++) {
-        Vital_Handicap[ix][0] = 7;
-        Vital_Handicap[ix][1] = 7;
-    }
-
-    VS_Stage = 0x14;
-    Order[0x8A] = 4;
-    Order_Timer[0x8A] = 1;
-
-    for (ix = 0; ix < 4; ix++) {
-        Message_Data[ix].order = 3;
-    }
-
-    effect_57_init(0x64, MENU_HEADER_MODE_MENU, 0, 0x3F, 2);
-    Order[0x64] = 1;
-    Order_Dir[0x64] = 8;
-    Order_Timer[0x64] = 1;
-    Menu_Suicide[0] = 0;
-    effect_04_init(0, 0, 0, 0x48);
-
-    for (ix = 0; ix < loop_counter; ix++) {
-        effect_61_init(0, ix + 0x50, 0, 0, (u32)ix, ix, 0x7047);
-        Order[ix + 0x50] = 1;
-        Order_Dir[ix + 0x50] = 4;
-        Order_Timer[ix + 0x50] = ix + 0x14;
-    }
-
-    Menu_Cursor_Move = loop_counter;
-}
-
-static bool should_skip_mode_selection(void) {
-    return Connect_Status == 0 && Menu_Cursor_Y[0] == 1;
-}
-
-static void handle_mode_select_input(struct _TASK* task_ptr, const s16 loop_counter) {
-    s16 PL_id;
-
-    if (should_skip_mode_selection()) {
-        Menu_Cursor_Y[0] = 2;
-    } else {
-        PL_id = 0;
-
-        if (MC_Move_Sub(Check_Menu_Lever(0, 0), 0, loop_counter - 1, 1) == 0) {
-            PL_id = 1;
-            MC_Move_Sub(Check_Menu_Lever(1, 0), 0, loop_counter - 1, 1);
-        }
-    }
-
-    switch (IO_Result) {
-    case 0x100:
-        switch (Menu_Cursor_Y[0]) {
-        case 0:
-            G_No[2] += 1;
-            Mode_Type = MODE_ARCADE;
-            task_ptr->r_no[0] = 5;
-            cpExitTask(TASK_SAVER);
-            Decide_PL(PL_id);
-            break;
-
-        case 1:
-            Setup_VS_Mode(task_ptr);
-            G_No[1] = 12;
-            G_No[2] = 1;
-            Mode_Type = MODE_VERSUS;
-            cpExitTask(TASK_MENU);
-            break;
-
-        case 2:
-        case 3:
-        case 4:
-        case 5:
-        case 6:
-            task_ptr->r_no[2] += 1;
-            task_ptr->free[0] = 0;
-            task_ptr->free[1] = Menu_Cursor_Y[0] + 2;
-            break;
-
-        default:
-            break;
-        }
-
-        SE_selected();
-        break;
-    }
-}
-
-static void handle_mode_select_load(struct _TASK* task_ptr) {
-    if (task_ptr->free[3]) {
-        FadeOut(1, 0xFF, 8);
-
-        if (SaveMove() > 0) {
-            return;
-        }
-
-        task_ptr->free[3] = 0;
-        Forbid_Reset = 0;
-    }
-
-    if (Menu_Sub_case1(task_ptr) != 0) {
-        Order[0x4E] = 2;
-        Order_Dir[0x4E] = 0;
-        Order_Timer[0x4E] = 1;
-        checkAdxFileLoaded();
-        checkSelObjFileLoaded();
-    }
-}
-
-void Mode_Select(struct _TASK* task_ptr) {
-    const bool supports_exit = App_SupportsExit();
-    const s16 loop_counter = supports_exit ? 7 : 6;
-
-    switch (task_ptr->r_no[2]) {
-    case 0:
-        initialize_mode_select(task_ptr, loop_counter);
-        break;
-
-    case 1:
-        handle_mode_select_load(task_ptr);
-        break;
-
-    case 2:
-        if (FadeIn(1, 0x19, 8) != 0) {
-            task_ptr->r_no[2] += 1;
-            Suicide[3] = 0;
-        }
-
-        break;
-
-    case 3:
-        handle_mode_select_input(task_ptr, loop_counter);
-        break;
-
-    default:
-        Exit_Sub(task_ptr, 0, task_ptr->free[1]);
-        break;
-    }
-}
-
-void Setup_VS_Mode(struct _TASK* task_ptr) {
-    task_ptr->r_no[0] = 5;
-    cpExitTask(TASK_SAVER);
-    plw[0].wu.operator = 1;
-    plw[1].wu.operator = 1;
-    Operator_Status[0] = 1;
-    Operator_Status[1] = 1;
-    grade_check_work_1st_init(0, 0);
-    grade_check_work_1st_init(0, 1);
-    grade_check_work_1st_init(1, 0);
-    grade_check_work_1st_init(1, 1);
-    Setup_Training_Difficulty();
-}
-
-void Menu_in_Sub(struct _TASK* task_ptr) {
-    FadeOut(1, 0xFF, 8);
-    task_ptr->r_no[2] += 1;
-    task_ptr->timer = 5;
-    Menu_Common_Init();
-    Menu_Cursor_Y[0] = Cursor_Y_Pos[0][1];
-    Menu_Suicide[0] = 1;
-    Menu_Suicide[1] = 0;
-    Order[0x64] = 4;
-    Order_Timer[0x64] = 1;
-}
-
-static void handle_select_game_input(struct _TASK* task_ptr) {
-    u16 sw;
-
-    imgSelectGameButton();
-    sw = (~plsw_01[0] & plsw_00[0]) | (~plsw_01[1] & plsw_00[1]); // potential macro
-    sw &= (SWK_SOUTH | SWK_EAST);
-
-    if (sw != 0) {
-        if (sw != (SWK_SOUTH | SWK_EAST)) {
-            if (sw & SWK_SOUTH) {
-                task_ptr->free[0] = 1;
-            }
-
-            SE_selected();
-            FadeInit();
-            task_ptr->r_no[2] = 8;
-        }
-    }
-}
-
-static void handle_select_game_fade_out(struct _TASK* task_ptr) {
-    imgSelectGameButton();
-
-    if (FadeOut(1, 25, 8) != 0) {
-        if (task_ptr->free[0]) {
-            task_ptr->r_no[2] = 10;
-            sound_all_off();
-        } else {
-            task_ptr->r_no[2] = 9;
-        }
-    }
-}
-
-void toSelectGame(struct _TASK* task_ptr) {
-    switch (task_ptr->r_no[2]) {
-    case 0:
-        Forbid_Reset = 1;
-        Menu_in_Sub(task_ptr);
-        Setup_BG(1, 0x200, 0);
-        effect_66_init(0x8A, 8, 1, 0, -1, -1, -0x7FF2);
-        Order[0x8A] = 3;
-        Order_Timer[0x8A] = 1;
-        task_ptr->free[0] = 0;
-        task_ptr->timer = 0x10;
-        break;
-
-    case 1:
-        if (Menu_Sub_case1(task_ptr) != 0) {
-            Message_Data->kind_req = 5;
-            Message_Data->request = 0;
-            Message_Data->order = 1;
-            Message_Data->timer = 2;
-            Message_Data->pos_x = 0;
-            Message_Data->pos_y = 0xA0;
-            Message_Data->pos_z = 0x18;
-            effect_45_init(0, 0, 2);
-        }
-
-        break;
-
-    case 2:
-        if (FadeIn(1, 0x19, 8) != 0) {
-            task_ptr->r_no[2] += 1;
-        }
-
-        imgSelectGameButton();
-        break;
-
-    case 3:
-        handle_select_game_input(task_ptr);
-        break;
-
-    case 8:
-        handle_select_game_fade_out(task_ptr);
-        break;
-
-    case 9:
-        Menu_Suicide[0] = 0;
-        Menu_Suicide[1] = 1;
-        task_ptr->r_no[1] = 1;
-        task_ptr->r_no[2] = 0;
-        task_ptr->r_no[3] = 0;
-        task_ptr->free[0] = 0;
-        FadeOut(1, 0xFF, 8);
-        Forbid_Reset = 0;
-        break;
-
-    case 10:
-        Exit_sound_system();
-        task_ptr->r_no[2] += 1;
-        FadeOut(1, 0xFF, 8);
-        break;
-
-    default:
-        App_Exit();
-        FadeOut(1, 0xFF, 8);
-        break;
-    }
-}
-
-void imgSelectGameButton() {
-    dispButtonImage2(0x74, 0x6B, 0x18, 0x20, 0x1A, 0, 4, 0);
-    dispButtonImage2(0xB2, 0x6B, 0x18, 0x20, 0x1A, 0, 5, 0);
-}
-
-static bool is_option_selection_input(void) {
-    if (MC_Move_Sub(Check_Menu_Lever(0, 0), 0, 5, 0xFF) == 0) {
-        MC_Move_Sub(Check_Menu_Lever(1, 0), 0, 5, 0xFF);
-    }
-
-    switch (IO_Result) {
-    case 0x100:
-    case 0x200:
-        return true;
-
-    default:
-        return false;
-    }
-}
-
-void Option_Select(struct _TASK* task_ptr) {
-    s16 ix;
-    static const s16 option_items[6] = { 7, 8, 9, 10, 12, 13 };
-    static const s16 option_routines[6] = { 9, 10, 11, 12, 14, 15 };
-
-    switch (task_ptr->r_no[2]) {
-    case 0:
-        Menu_in_Sub(task_ptr);
-        Order[0x4E] = 2;
-        Order_Dir[0x4E] = 0;
-        Order_Timer[0x4E] = 1;
-        effect_57_init(0x4F, MENU_HEADER_OPTION_MENU, 0, 0x3F, 2);
-        Order[0x4F] = 1;
-        Order_Dir[0x4F] = 8;
-        Order_Timer[0x4F] = 1;
-        effect_04_init(1, 1, 0, 0x48);
-
-        if (Menu_Cursor_Y[0] >= 6) {
-            Menu_Cursor_Y[0] = 5;
-        }
-
-        for (ix = 0; ix < 6; ix++) {
-            effect_61_init(0, ix + 0x50, 0, 1, option_items[ix], ix, 0x7047);
-            Order[ix + 0x50] = 1;
-            Order_Dir[ix + 0x50] = 4;
-            Order_Timer[ix + 0x50] = ix + 0x14;
-        }
-
-        Menu_Cursor_Move = 6;
-        break;
-
-    case 1:
-        if (Menu_Sub_case1(task_ptr) != 0) {
-            checkSelObjFileLoaded();
-        }
-
-        break;
-
-    case 2:
-        if (FadeIn(1, 0x19, 8) != 0) {
-            task_ptr->r_no[2] += 1;
-            Suicide[3] = 0;
-        }
-
-        break;
-
-    case 3:
-        if (!is_option_selection_input()) {
-            return;
-        }
-
-        SE_selected();
-
-        if (Menu_Cursor_Y[0] == 5 || IO_Result == 0x200) {
-            Menu_Suicide[0] = 0;
-            Menu_Suicide[1] = 1;
-            task_ptr->r_no[1] = 1;
-            task_ptr->r_no[2] = 0;
-            task_ptr->r_no[3] = 0;
-            task_ptr->free[0] = 0;
-            Order[0x4F] = 4;
-            Order_Timer[0x4F] = 4;
-
-            if (Check_Change_Contents()) {
-                SaveInit(SAVE_FILE_SETTINGS, SAVE_MODE_SAVE);
-                task_ptr->free[3] = 1;
-                Forbid_Reset = 1;
-                Copy_Check_w();
-            }
-
-            break;
-        }
-
-        task_ptr->r_no[2] += 1;
-        task_ptr->free[0] = 0;
-        X_Adjust_Buff[0] = X_Adjust;
-        X_Adjust_Buff[1] = X_Adjust;
-        X_Adjust_Buff[2] = X_Adjust;
-        Y_Adjust_Buff[0] = Y_Adjust;
-        Y_Adjust_Buff[1] = Y_Adjust;
-        Y_Adjust_Buff[2] = Y_Adjust;
-        break;
-
-    default:
-        Exit_Sub(task_ptr, 1, option_routines[Menu_Cursor_Y[0]]);
-        break;
-    }
-}
 
 const u8 Setup_Index_64[10] = { 1, 2, 3, 3, 4, 5, 6, 7, 8, 8 };
 
@@ -819,6 +412,43 @@ static void menu_select_confirm(struct _TASK* task_ptr) {
     }
 }
 
+static void initialize_pause_menu(struct _TASK* task_ptr) {
+    task_ptr->r_no[2]++;
+    Menu_Common_Init();
+    Menu_Cursor_Y[0] = Cursor_Y_Pos[0][0];
+    Menu_Suicide[0] = 0;
+    Menu_Suicide[1] = 0;
+    Menu_Suicide[2] = 0;
+    effect_10_init(0, 0, 0, 0, 0, 0x14, 0xC);
+    effect_10_init(0, 0, 2, 2, 0, 0x16, 0x10);
+    switch (Mode_Type) {
+    case MODE_VERSUS:
+        effect_10_init(0, 0, 1, 5, 0, 0x10, 0xE);
+        break;
+    case MODE_REPLAY:
+        effect_10_init(0, 0, 1, 4, 0, 0x15, 0xE);
+        break;
+    default:
+        effect_10_init(0, 0, 1, 1, 0, 0x11, 0xE);
+        break;
+    }
+}
+
+static void handle_pause_menu_input(struct _TASK* task_ptr) {
+    IO_Result = MC_Move_Sub(Check_Menu_Lever(Pause_ID, 0), 0, 2, 0xFF);
+    switch (IO_Result) {
+    case SWK_START:
+    case SWK_EAST:
+        task_ptr->r_no[2] = 99;
+        Exit_Menu = 1;
+        SE_selected();
+        break;
+    case SWK_SOUTH:
+        menu_select_confirm(task_ptr);
+        break;
+    }
+}
+
 void Menu_Select(struct _TASK* task_ptr) {
     if (Check_Pad_in_Pause(task_ptr) != 0) {
         return;
@@ -831,47 +461,11 @@ void Menu_Select(struct _TASK* task_ptr) {
         /* fallthrough */
 
     case 1:
-        task_ptr->r_no[2]++;
-        Menu_Common_Init();
-        Menu_Cursor_Y[0] = Cursor_Y_Pos[0][0];
-        Menu_Suicide[0] = 0;
-        Menu_Suicide[1] = 0;
-        Menu_Suicide[2] = 0;
-        effect_10_init(0, 0, 0, 0, 0, 0x14, 0xC);
-        effect_10_init(0, 0, 2, 2, 0, 0x16, 0x10);
-
-        switch (Mode_Type) {
-        case MODE_VERSUS:
-            effect_10_init(0, 0, 1, 5, 0, 0x10, 0xE);
-            break;
-
-        case MODE_REPLAY:
-            effect_10_init(0, 0, 1, 4, 0, 0x15, 0xE);
-            break;
-
-        default:
-            effect_10_init(0, 0, 1, 1, 0, 0x11, 0xE);
-            break;
-        }
-
+        initialize_pause_menu(task_ptr);
         break;
 
     case 2:
-        IO_Result = MC_Move_Sub(Check_Menu_Lever(Pause_ID, 0), 0, 2, 0xFF);
-
-        switch (IO_Result) {
-        case SWK_START:
-        case SWK_EAST:
-            task_ptr->r_no[2] = 99;
-            Exit_Menu = 1;
-            SE_selected();
-            break;
-
-        case SWK_SOUTH:
-            menu_select_confirm(task_ptr);
-            break;
-        }
-
+        handle_pause_menu_input(task_ptr);
         break;
 
     case 3:
@@ -880,44 +474,48 @@ void Menu_Select(struct _TASK* task_ptr) {
     }
 }
 
+static void move_yes_no_cursor_up(void) {
+    Menu_Cursor_Y[0]--;
+    if (Menu_Cursor_Y[0] < 0) {
+        Menu_Cursor_Y[0] = 0;
+    } else {
+        SE_dir_cursor_move();
+    }
+}
+
+static void move_yes_no_cursor_down(void) {
+    Menu_Cursor_Y[0]++;
+    if (Menu_Cursor_Y[0] > 1) {
+        Menu_Cursor_Y[0] = 1;
+    } else {
+        SE_dir_cursor_move();
+    }
+}
+
+static s32 confirm_yes_no_cursor(struct _TASK* task_ptr, u16 sw) {
+    if (Menu_Cursor_Y[0] || sw == 0x200) {
+        task_ptr->r_no[2] = 1;
+        Menu_Suicide[0] = 0;
+        Menu_Suicide[1] = 1;
+        Cursor_Y_Pos[0][0] = 2;
+        return 1;
+    }
+    Soft_Reset_Sub();
+    return -1;
+}
+
 s32 Yes_No_Cursor_Move_Sub(struct _TASK* task_ptr) {
     u16 sw = ~(plsw_01[Pause_ID]) & plsw_00[Pause_ID];
-
     switch (sw) {
     case 0x4:
-        Menu_Cursor_Y[0]--;
-
-        if (Menu_Cursor_Y[0] < 0) {
-            Menu_Cursor_Y[0] = 0;
-        } else {
-            SE_dir_cursor_move();
-        }
-
+        move_yes_no_cursor_up();
         break;
-
     case 0x8:
-        Menu_Cursor_Y[0]++;
-
-        if (Menu_Cursor_Y[0] > 1) {
-            Menu_Cursor_Y[0] = 1;
-        } else {
-            SE_dir_cursor_move();
-        }
-
+        move_yes_no_cursor_down();
         break;
-
     case 0x200:
     case 0x100:
-        if (Menu_Cursor_Y[0] || sw == 0x200) {
-            task_ptr->r_no[2] = 1;
-            Menu_Suicide[0] = 0;
-            Menu_Suicide[1] = 1;
-            Cursor_Y_Pos[0][0] = 2;
-            return 1;
-        }
-
-        Soft_Reset_Sub();
-        return -1;
+        return confirm_yes_no_cursor(task_ptr, sw);
     }
 
     return 0;
@@ -1087,77 +685,78 @@ void bg_etc_write_ex(s16 type) {
     base_y_pos = 40;
 }
 
-void Wait_Load_Save(struct _TASK* task_ptr) {
+static void begin_load_save_exit(struct _TASK* task_ptr) {
+    if (vm_w.Request != 0) {
+        return;
+    }
+    task_ptr->free[0] = 0;
+    task_ptr->free[1]++;
+    if (task_ptr->r_no[1] == 5) {
+        task_ptr->free[2] = 18;
+    } else {
+        task_ptr->free[2] = task_ptr->r_no[1];
+    }
+    Exit_Sub(task_ptr, 2, task_ptr->free[2]);
+}
+
+static void set_load_save_exit_order(struct _TASK* task_ptr, s16* order) {
+    switch (task_ptr->r_no[1]) {
+    case 13:
+        *order = 105;
+        break;
+    case 17:
+        task_ptr->r_no[2] = 99;
+        /* fallthrough */
+    case 6:
+        *order = 110;
+        break;
+    case 19:
+    case 20:
+        *order = 112;
+        break;
+    case 23:
+        task_ptr->r_no[0] = 0;
+        task_ptr->r_no[2] = 99;
+        task_ptr->free[0] = 1;
+        task_ptr->free[1] = 8;
+        *order = 105;
+        break;
+    }
+}
+
+static void finish_load_save_exit(struct _TASK* task_ptr) {
     s16 ix;
 
+    if (!Exit_Sub(task_ptr, 2, task_ptr->free[2])) {
+        return;
+    }
+    task_ptr->free[1]++;
+    task_ptr->timer = 1;
+    for (ix = 0; ix < 4; ix++) {
+        Menu_Suicide[ix] = 1;
+    }
+    set_load_save_exit_order(task_ptr, &ix);
+    Order[ix] = 4;
+    Order_Timer[ix] = 1;
+}
+
+static void complete_load_save_wait(struct _TASK* task_ptr) {
+    FadeOut(1, 0xFF, 8);
+    if (--task_ptr->timer == 0) {
+        task_ptr->r_no[0] = 0;
+    }
+}
+
+void Wait_Load_Save(struct _TASK* task_ptr) {
     switch (task_ptr->free[1]) {
     case 0:
-        if (vm_w.Request != 0) {
-            break;
-        }
-
-        task_ptr->free[0] = 0;
-        task_ptr->free[1]++;
-
-        if (task_ptr->r_no[1] == 5) {
-            task_ptr->free[2] = 18;
-        } else {
-            task_ptr->free[2] = task_ptr->r_no[1];
-        }
-
-        Exit_Sub(task_ptr, 2, task_ptr->free[2]);
+        begin_load_save_exit(task_ptr);
         break;
-
     case 1:
-        if (!Exit_Sub(task_ptr, 2, task_ptr->free[2])) {
-            break;
-        }
-
-        task_ptr->free[1]++;
-        task_ptr->timer = 1;
-
-        for (ix = 0; ix < 4; ix++) {
-            Menu_Suicide[ix] = 1;
-        }
-
-        switch (task_ptr->r_no[1]) {
-        case 13:
-            ix = 105;
-            break;
-
-        case 17:
-            task_ptr->r_no[2] = 99;
-            /* fallthrough */
-
-        case 6:
-            ix = 110;
-            break;
-
-        case 19:
-        case 20:
-            ix = 112;
-            break;
-
-        case 23:
-            ix = 105;
-            task_ptr->r_no[0] = 0;
-            task_ptr->r_no[2] = 99;
-            task_ptr->free[0] = 1;
-            task_ptr->free[1] = 8;
-            break;
-        }
-
-        Order[ix] = 4;
-        Order_Timer[ix] = 1;
+        finish_load_save_exit(task_ptr);
         break;
-
     case 2:
-        FadeOut(1, 0xFF, 8);
-
-        if (--task_ptr->timer == 0) {
-            task_ptr->r_no[0] = 0;
-        }
-
+        complete_load_save_wait(task_ptr);
         break;
     }
 }
@@ -1273,6 +872,25 @@ static bool is_disallowed_dummy_selection(s16 id, s16 type, s16 cursor_id) {
            Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] == 4;
 }
 
+static void adjust_dummy_option(u16 sw, s16 id, s16 type, s16 cursor_id) {
+    s16 max = Menu_Max_Data_Tr[id][type][Menu_Cursor_Y[cursor_id]];
+    if (sw == SWK_WEST) {
+        Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]]--;
+    } else {
+        Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]]++;
+    }
+    if (Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] < 0) {
+        Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] = max;
+    }
+    if (Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] > max) {
+        Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] = 0;
+    }
+    if (is_disallowed_dummy_selection(id, type, cursor_id)) {
+        Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] = sw == SWK_WEST ? 3 : 0;
+    }
+    SE_dir_cursor_move();
+}
+
 void Dummy_Move_Sub_LR(u16 sw, s16 id, s16 type, s16 cursor_id) {
     s16 max = Menu_Max_Data_Tr[id][type][Menu_Cursor_Y[cursor_id]];
 
@@ -1282,31 +900,8 @@ void Dummy_Move_Sub_LR(u16 sw, s16 id, s16 type, s16 cursor_id) {
 
     switch (sw) {
     case 4:
-        Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]]--;
-
-        if (Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] < 0) {
-            Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] = max;
-        }
-
-        if (is_disallowed_dummy_selection(id, type, cursor_id)) {
-            Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] = 3;
-        }
-
-        SE_dir_cursor_move();
-        break;
-
     case 8:
-        Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]]++;
-
-        if (Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] > max) {
-            Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] = 0;
-        }
-
-        if (is_disallowed_dummy_selection(id, type, cursor_id)) {
-            Training[2].contents[id][type][Menu_Cursor_Y[cursor_id]] = 0;
-        }
-
-        SE_dir_cursor_move();
+        adjust_dummy_option(sw, id, type, cursor_id);
         break;
 
     default:
@@ -1318,62 +913,62 @@ void Dummy_Move_Sub_LR(u16 sw, s16 id, s16 type, s16 cursor_id) {
     }
 }
 
+static void prepare_recording_training(void) {
+    Record_Data_Tr = 1;
+    Training[0] = Training[2];
+    Training[0].contents[1][0][2] = 1;
+    Training[1] = Training[2];
+    switch (Training[0].contents[1][0][0]) {
+    case 0:
+        control_pl_rno = 0;
+        break;
+    case 1:
+        control_pl_rno = 1;
+        break;
+    case 2:
+        control_pl_rno = 2;
+        break;
+    }
+}
+
+static void start_selected_training(struct _TASK* task_ptr) {
+    if (Menu_Cursor_Y[0] == 0) {
+            Play_Mode = 1;
+    } else {
+            Play_Mode = 3;
+    }
+    All_Clear_Timer();
+    Check_Replay();
+    if (Menu_Cursor_Y[0] == 1) {
+        Replay_Status[Training_ID] = 0;
+        Replay_Status[Training_ID ^ 1] = 3;
+        Training[0] = Training[1];
+        Training[0].contents[1][0][2] = Training[2].contents[1][0][2];
+        Training[0].contents[1][0][3] = Training[2].contents[1][0][3];
+        control_pl_rno = 99;
+    }
+    task_ptr->r_no[0] = 10;
+    task_ptr->r_no[1] = 0;
+    task_ptr->r_no[2] = 0;
+    task_ptr->r_no[3] = 0;
+    Menu_Suicide[0] = 1;
+    Game_pause = 0;
+    Pause_Down = 0;
+    save_w[Present_Mode].Time_Limit = 60;
+    count_cont_init(0);
+    Training[0].contents[1][1][3] = Menu_Cursor_Y[0];
+    init_omop();
+    set_init_A4_flag();
+    Training_Cursor = Menu_Cursor_Y[0];
+}
+
 void handle_blocking_training_selection(struct _TASK* task_ptr) {
     switch (Menu_Cursor_Y[0]) {
     case 0:
-        Record_Data_Tr = 1;
-        Training[0] = Training[2];
-        Training[0].contents[1][0][2] = 1;
-        Training[1] = Training[2];
-
-        switch (Training[0].contents[1][0][0]) {
-        case 0:
-            control_pl_rno = 0;
-            break;
-
-        case 1:
-            control_pl_rno = 1;
-            break;
-
-        case 2:
-            control_pl_rno = 2;
-            break;
-        }
-
+        prepare_recording_training();
         /* fallthrough */
-
     case 1:
-        if (Menu_Cursor_Y[0] == 0) {
-            Play_Mode = 1;
-        } else {
-            Play_Mode = 3;
-        }
-
-        All_Clear_Timer();
-        Check_Replay();
-
-        if (Menu_Cursor_Y[0] == 1) {
-            Replay_Status[Training_ID] = 0;
-            Replay_Status[Training_ID ^ 1] = 3;
-            Training[0] = Training[1];
-            Training[0].contents[1][0][2] = Training[2].contents[1][0][2];
-            Training[0].contents[1][0][3] = Training[2].contents[1][0][3];
-            control_pl_rno = 99;
-        }
-
-        task_ptr->r_no[0] = 10;
-        task_ptr->r_no[1] = 0;
-        task_ptr->r_no[2] = 0;
-        task_ptr->r_no[3] = 0;
-        Menu_Suicide[0] = 1;
-        Game_pause = 0;
-        Pause_Down = 0;
-        save_w[Present_Mode].Time_Limit = 60;
-        count_cont_init(0);
-        Training[0].contents[1][1][3] = Menu_Cursor_Y[0];
-        init_omop();
-        set_init_A4_flag();
-        Training_Cursor = Menu_Cursor_Y[0];
+        start_selected_training(task_ptr);
         break;
 
     case 2:
@@ -1428,9 +1023,39 @@ void Training_Exit_Sub(struct _TASK* task_ptr) {
     effect_10_init(0, 1, 1, 1, 1, 26, 15);
 }
 
-void Character_Change(struct _TASK* task_ptr) {
+static void wait_for_character_change_load(struct _TASK* task_ptr) {
+    if ((task_ptr->timer -= 1) != 0) {
+        return;
+    }
+    if (!Check_LDREQ_Break()) {
+        task_ptr->r_no[2]++;
+        Switch_Screen_Init(0);
+        return;
+    }
+    task_ptr->timer = 1;
+}
+
+static void finish_character_change(struct _TASK* task_ptr) {
     s16 ix;
 
+    if (Switch_Screen(0) == 0) {
+        return;
+    }
+    task_ptr->r_no[2]++;
+    Cover_Timer = 0x17;
+    G_No[1] = 1;
+    G_No[2] = 0;
+    G_No[3] = 0;
+    for (ix = 0; ix < 2; ix++) {
+        Sel_PL_Complete[ix] = 0;
+        Sel_Arts_Complete[ix] = 0;
+        plw[ix].wu.operator = 1;
+        Operator_Status[ix] = 1;
+    }
+    cpExitTask(TASK_MENU);
+}
+
+void Character_Change(struct _TASK* task_ptr) {
     if (!(Check_Pad_in_Pause(task_ptr) == 0)) {
         return;
     }
@@ -1443,43 +1068,30 @@ void Character_Change(struct _TASK* task_ptr) {
         break;
 
     case 1:
-        if ((task_ptr->timer -= 1) == 0) {
-            if (!Check_LDREQ_Break()) {
-                task_ptr->r_no[2]++;
-                Switch_Screen_Init(0);
-                return;
-            }
-
-            task_ptr->timer = 1;
-            return;
-        }
+        wait_for_character_change_load(task_ptr);
         break;
 
     case 2:
-        if (Switch_Screen(0) != 0) {
-            task_ptr->r_no[2]++;
-            Cover_Timer = 0x17;
-            G_No[1] = 1;
-            G_No[2] = 0;
-            G_No[3] = 0;
-
-            for (ix = 0; ix < 2; ix++) {
-                Sel_PL_Complete[ix] = 0;
-                Sel_Arts_Complete[ix] = 0;
-                plw[ix].wu.operator = 1;
-                Operator_Status[ix] = 1;
-            }
-
-            cpExitTask(TASK_MENU);
-        }
+        finish_character_change(task_ptr);
         break;
     }
 }
 
-void Default_Training_Data(s32 flag) {
+static void clear_training_data(void) {
     s16 ix;
     s16 ix2;
     s16 ix3;
+
+    for (ix = 0; ix < 2; ix++) {
+        for (ix2 = 0; ix2 < 2; ix2++) {
+            for (ix3 = 0; ix3 < 6; ix3++) {
+                Training[0].contents[ix][ix2][ix3] = 0;
+            }
+        }
+    }
+}
+
+void Default_Training_Data(s32 flag) {
 
     if (flag == 0) {
         if (!mpp_w.initTrainingData) {
@@ -1489,13 +1101,7 @@ void Default_Training_Data(s32 flag) {
         mpp_w.initTrainingData = false;
     }
 
-    for (ix = 0; ix < 2; ix++) {
-        for (ix2 = 0; ix2 < 2; ix2++) {
-            for (ix3 = 0; ix3 < 6; ix3++) {
-                Training[0].contents[ix][ix2][ix3] = 0;
-            }
-        }
-    }
+    clear_training_data();
 
     Training[0].contents[0][1][TRAINING_OPTION_DAMAGE] = save_w->Damage_Level;
     Training[0].contents[0][1][TRAINING_OPTION_DIFFICULTY] = save_w->Difficulty;
@@ -1554,299 +1160,6 @@ void Back_to_Mode_Select(struct _TASK* task_ptr) {
     BGM_Request_Code_Check(0x41);
 }
 
-static void update_extra_option_message() {
-    if (Menu_Cursor_Y[1] != Menu_Cursor_Y[0]) {
-        SE_cursor_move();
-        save_w[Present_Mode].extra_option.contents[Menu_Page][Menu_Max] = 1;
-
-        if (Menu_Cursor_Y[0] < Menu_Max) {
-            Message_Data->order = 1;
-            Message_Data->request = Ex_Account_Data[Menu_Page] + Menu_Cursor_Y[0];
-            Message_Data->timer = 2;
-
-            if (msgExtraTbl[0]->msgNum[Menu_Cursor_Y[0] + (Menu_Page * 8)] == 1) {
-                Message_Data->pos_y = 54;
-            } else {
-                Message_Data->pos_y = 62;
-            }
-        } else {
-            Message_Data->order = 1;
-            Message_Data->request = save_w[Present_Mode].extra_option.contents[Menu_Page][Menu_Max] + 32;
-            Message_Data->timer = 2;
-            Message_Data->pos_y = 54;
-        }
-    }
-}
-
-static void handle_extra_option_page_action(struct _TASK* task_ptr) {
-    switch (save_w[Present_Mode].extra_option.contents[Menu_Page][Menu_Max]) {
-    case 0:
-        task_ptr->r_no[2] = 1;
-        task_ptr->timer = 5;
-
-        if (--Menu_Page < 0) {
-            Menu_Page = Page_Max;
-        }
-
-        break;
-
-    case 2:
-        task_ptr->r_no[2] = 1;
-        task_ptr->timer = 5;
-
-        if (++Menu_Page > Page_Max) {
-            Menu_Page = 0;
-        }
-
-        break;
-
-    default:
-        Return_Option_Mode_Sub(task_ptr);
-        save_w[4].extra_option = save_w[1].extra_option;
-        save_w[5].extra_option = save_w[1].extra_option;
-        Order[115] = 4;
-        Order_Timer[115] = 4;
-        break;
-    }
-
-    SE_selected();
-}
-
-static void select_previous_extra_option_page(struct _TASK* task_ptr) {
-    task_ptr->r_no[2] = 1;
-    task_ptr->timer = 5;
-
-    if (--Menu_Page < 0) {
-        Menu_Page = Page_Max;
-    }
-
-    SE_dir_selected();
-}
-
-static void select_next_extra_option_page(struct _TASK* task_ptr) {
-    task_ptr->r_no[2] = 1;
-    task_ptr->timer = 5;
-
-    if (++Menu_Page > Page_Max) {
-        Menu_Page = 0;
-    }
-
-    SE_dir_selected();
-}
-
-static void read_second_extra_option_input(struct _TASK* task_ptr) {
-    if (IO_Result == 0) {
-        Pause_ID = 1;
-        Dir_Move_Sub(task_ptr, 1);
-    }
-}
-
-static void reset_extra_options(void) {
-    save_w[Present_Mode].extra_option = save_w[0].extra_option;
-    SE_selected();
-}
-
-static bool should_reset_extra_options(void) {
-    return Menu_Page == 0 && Menu_Cursor_Y[0] == 6;
-}
-
-static bool should_handle_extra_option_page_action(void) {
-    return Menu_Cursor_Y[0] == Menu_Max;
-}
-
-static void handle_extra_option_selection(struct _TASK* task_ptr) {
-    if (should_reset_extra_options()) {
-        reset_extra_options();
-        return;
-    }
-
-    if (!should_handle_extra_option_page_action()) {
-        return;
-    }
-
-    handle_extra_option_page_action(task_ptr);
-}
-
-static void advance_extra_option_fade(struct _TASK* task_ptr) {
-    if (--task_ptr->timer == 0) {
-        task_ptr->r_no[2]++;
-        task_ptr->r_no[3] = 1;
-        FadeInit();
-    }
-}
-
-static void handle_extra_option_input(struct _TASK* task_ptr) {
-    Pause_ID = 0;
-    Dir_Move_Sub(task_ptr, 0);
-    read_second_extra_option_input(task_ptr);
-
-    update_extra_option_message();
-
-    switch (IO_Result) {
-    case 0x200:
-        Return_Option_Mode_Sub(task_ptr);
-        Order[115] = 4;
-        Order_Timer[115] = 4;
-        save_w[4].extra_option = save_w[1].extra_option;
-        save_w[5].extra_option = save_w[1].extra_option;
-        SE_dir_selected();
-        break;
-
-    case 0x80:
-    case 0x800:
-        select_previous_extra_option_page(task_ptr);
-        break;
-
-    case 0x40:
-    case 0x400:
-        select_next_extra_option_page(task_ptr);
-        break;
-
-    case 0x100:
-        handle_extra_option_selection(task_ptr);
-        break;
-    }
-}
-
-void Extra_Option(struct _TASK* task_ptr) {
-    Menu_Cursor_Y[1] = Menu_Cursor_Y[0];
-
-    switch (task_ptr->r_no[2]) {
-    case 0:
-        FadeOut(1, 0xFF, 8);
-        task_ptr->r_no[2]++;
-        task_ptr->r_no[3] = 0;
-        task_ptr->timer = 5;
-        Menu_Suicide[1] = 1;
-        Menu_Suicide[2] = 0;
-        Menu_Page = 0;
-        Page_Max = 3;
-        Menu_Page_Buff = Menu_Page;
-        Message_Data->kind_req = 4;
-        break;
-
-    case 1:
-        FadeOut(1, 0xFF, 8);
-        task_ptr->r_no[2]++;
-        Setup_Next_Page(task_ptr, task_ptr->r_no[3]);
-        /* fallthrough */
-
-    case 2:
-        FadeOut(1, 0xFF, 8);
-        advance_extra_option_fade(task_ptr);
-
-        break;
-
-    case 3:
-        if (FadeIn(1, 25, 8)) {
-            task_ptr->r_no[2]++;
-            break;
-        }
-
-        break;
-
-    case 4:
-        handle_extra_option_input(task_ptr);
-        break;
-    }
-}
-
-static bool is_extra_option_page_item(void) {
-    return Menu_Page_Buff != 0 || Menu_Cursor_Y[0] != 4;
-}
-
-static void move_extra_option_left(u8 last_pos) {
-    if (is_extra_option_page_item()) {
-        SE_dir_cursor_move();
-    }
-
-    save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]]--;
-
-    if (Menu_Cursor_Y[0] == Menu_Max) {
-        if (save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] < 0) {
-            save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] = 0;
-            IO_Result = 0x80;
-            return;
-        }
-
-        if (save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] != last_pos) {
-            Message_Data->order = 1;
-            Message_Data->request = save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Max] + 32;
-            Message_Data->timer = 2;
-        }
-    } else if (save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] < 0) {
-        save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] =
-            Ex_Menu_Max_Data[Menu_Page][Menu_Cursor_Y[0]];
-    }
-}
-
-static void move_extra_option_right(u8 last_pos) {
-    if (is_extra_option_page_item()) {
-        SE_dir_cursor_move();
-    }
-
-    save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]]++;
-
-    if (Menu_Cursor_Y[0] == Menu_Max) {
-        if (save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] > 2) {
-            save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] = 2;
-            IO_Result = 0x400;
-            return;
-        }
-
-        if (save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] > 2) {
-            save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] = 2;
-        }
-
-        if (save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] != last_pos) {
-            Message_Data->order = 1;
-            Message_Data->request = save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Max] + 32;
-            Message_Data->timer = 2;
-        }
-    } else if (
-        save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] >
-        Ex_Menu_Max_Data[Menu_Page][Menu_Cursor_Y[0]]
-    ) {
-        save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] = 0;
-    }
-}
-
-void Ex_Move_Sub_LR(u16 sw, s16 PL_id) {
-    u8 last_pos = save_w[Present_Mode].extra_option.contents[Menu_Page][Menu_Cursor_Y[0]];
-
-    switch (sw) {
-    case 4:
-        move_extra_option_left(last_pos);
-        return;
-
-    case 8:
-        move_extra_option_right(last_pos);
-        return;
-
-    case 0x400:
-        if (Interface_Type[PL_id] == 2) {
-            break;
-        }
-
-    case 0x100:
-        if (is_extra_option_page_item()) {
-            SE_dir_cursor_move();
-        }
-
-        if (Menu_Cursor_Y[0] == Menu_Max) {
-            break;
-        }
-
-        save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]]++;
-
-        if (save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] >
-            Ex_Menu_Max_Data[Menu_Page][Menu_Cursor_Y[0]]) {
-            save_w[1].extra_option.contents[Menu_Page_Buff][Menu_Cursor_Y[0]] = 0;
-        }
-
-        return;
-    }
-}
 
 static void handle_end_replay_selection(struct _TASK* task_ptr) {
     s16 ix;
@@ -1880,38 +1193,55 @@ static void handle_end_replay_selection(struct _TASK* task_ptr) {
     }
 }
 
-void End_Replay_Menu(struct _TASK* task_ptr) {
-    s16 ix;
-    s16 ans;
+static bool initialize_end_replay_menu(struct _TASK* task_ptr) {
+    if (Allow_a_battle_f == 0) {
+        return false;
+    }
+    task_ptr->r_no[1] += 1;
+    Pause_ID = Decide_ID;
+    Pause_Down = 1;
+    Game_pause = 0x81;
+    effect_A3_init(1, 0x16, 0x63, 0, 3, 0x82, 0x48, 1);
+    effect_A3_init(1, 0x16, 0x63, 1, 3, 0x88, 0x58, 1);
+    Order[0x8A] = 3;
+    Order_Timer[0x8A] = 1;
+    effect_66_init(0x8A, 0xA, 2, 7, -1, -1, -0x3FF6);
+    return true;
+}
 
+static void setup_end_replay_menu(struct _TASK* task_ptr) {
+    s16 ix;
+
+    task_ptr->r_no[1] += 1;
+    Menu_Common_Init();
+    Menu_Cursor_Y[0] = 0;
+    for (ix = 0; ix < 4; ix++) {
+        Menu_Suicide[ix] = 0;
+    }
+    effect_10_init(0, 0, 0, 4, 0, 0x14, 0xE);
+    effect_10_init(0, 6, 1, 2, 0, 0x16, 0x10);
+}
+
+static void handle_end_replay_confirmation(struct _TASK* task_ptr) {
+    switch (Yes_No_Cursor_Move_Sub(task_ptr)) {
+    case 1:
+        task_ptr->r_no[1] = 1;
+        break;
+    case -1:
+        Menu_Suicide[3] = 1;
+        break;
+    }
+}
+
+void End_Replay_Menu(struct _TASK* task_ptr) {
     switch (task_ptr->r_no[1]) {
     case 0:
-        if (Allow_a_battle_f == 0) {
+        if (!initialize_end_replay_menu(task_ptr)) {
             break;
         }
-
-        task_ptr->r_no[1] += 1;
-        Pause_ID = Decide_ID;
-        Pause_Down = 1;
-        Game_pause = 0x81;
-        effect_A3_init(1, 0x16, 0x63, 0, 3, 0x82, 0x48, 1);
-        effect_A3_init(1, 0x16, 0x63, 1, 3, 0x88, 0x58, 1);
-        Order[0x8A] = 3;
-        Order_Timer[0x8A] = 1;
-        effect_66_init(0x8A, 0xA, 2, 7, -1, -1, -0x3FF6);
         /* fallthrough */
-
     case 1:
-        task_ptr->r_no[1] += 1;
-        Menu_Common_Init();
-        Menu_Cursor_Y[0] = 0;
-
-        for (ix = 0; ix < 4; ix++) {
-            Menu_Suicide[ix] = 0;
-        }
-
-        effect_10_init(0, 0, 0, 4, 0, 0x14, 0xE);
-        effect_10_init(0, 6, 1, 2, 0, 0x16, 0x10);
+        setup_end_replay_menu(task_ptr);
         break;
 
     case 2:
@@ -1920,18 +1250,7 @@ void End_Replay_Menu(struct _TASK* task_ptr) {
         break;
 
     case 3:
-        ans = Yes_No_Cursor_Move_Sub(task_ptr);
-
-        switch (ans) {
-        case 1:
-            task_ptr->r_no[1] = 1;
-            break;
-
-        case -1:
-            Menu_Suicide[3] = 1;
-            break;
-        }
-
+        handle_end_replay_confirmation(task_ptr);
         break;
     }
 }
